# openapi.yaml
...
openapi: 3.0.3
info:
  title: PrepChef API
  description: Commercial kitchen marketplace API
  version: 1.0.0
  contact:
    email: api@prepchef.com
  license:
    name: MIT
servers:
  - url: https://api.prepchef.com/v1
    description: Production
  - url: https://staging-api.prepchef.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /api/bookings:
    post:
      tags: [Bookings]
      summary: Create a booking
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid request
        '409':
          description: Kitchen not available
  /api/bookings/{id}:
    get:
      tags: [Bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Booking found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found
...
  /api/v1/platform/reviews/kitchens/{kitchen_id}:
    get:
      tags: [Reviews]
      summary: List reviews for a kitchen
      description: Cursor-paginated list of customer reviews for the specified kitchen.
      parameters:
        - name: kitchen_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Cursor from a previous response's pagination metadata.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Maximum number of reviews to return.
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '422':
          description: Invalid cursor supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CreateBookingRequest:
      type: object
      required:
        - kitchen_id
        - user_id
        - start_time
        - end_time
      properties:
        kitchen_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
    Booking:
      type: object
      properties:
        booking_id:
          type: string
          format: uuid
        kitchen_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, confirmed, cancelled, rejected]
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        target:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    ErrorResponse:
      type: object
      required: [request_id, error]
      properties:
        request_id:
          type: string
        error:
          $ref: '#/components/schemas/ErrorDetail'
    CursorPageMeta:
      type: object
      required: [limit, has_more]
      properties:
        cursor:
          type: string
          nullable: true
        next_cursor:
          type: string
          nullable: true
        limit:
          type: integer
          minimum: 1
          maximum: 100
        has_more:
          type: boolean
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rating:
          type: number
          format: float
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, rating, created_at]
    ReviewListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        pagination:
          $ref: '#/components/schemas/CursorPageMeta'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
      description: Stable key used to guarantee idempotent processing of POST and PUT requests.
security:
  - BearerAuth: []